cyberdudebivash_open_port_checker_pro/requirements.txt
pyqt6
python-nmap
fpdf
requests